Classes common to transformation and update moved out of transformation package
