+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/client/CaseClientServiceException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.client;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.client.HttpStatusCodeException;
