+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/client/model/request/DocumentUrl.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.request;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.client.model.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
